PLATFORM-4046| move CircuitBreaker/CircuitBreakerFactory into separate file
