+++ lib/Wikia/src/CircuitBreaker/CircuitBreaker.php
@@ -4,7 +4,6 @@ namespace Wikia\CircuitBreaker;
 
 use Wikia\Service\Helios\ClientException;
 use Wikia\Service\Helios\Exception;
-use Wikia\Util\Statistics\BernoulliTrial;
 
 class CircuitBreakerOpen extends ClientException {
 	/** @var string */
@@ -41,21 +40,4 @@ interface CircuitBreaker {
 	 * @return bool
 	 */
 	public function SetOperationStatus( string $name, bool $status );
-}
-
-class CircuitBreakerFactory {
-	/**
-	 * @param BernoulliTrial $logSampler
-	 * @return ExternalCircuitBreaker|NoopCircuitBreaker
-	 */
-	public static function GetCircuitBreaker( BernoulliTrial $logSampler ) {
-		global $wgCircuitBreakerType;
-
-		switch ( $wgCircuitBreakerType ) {
-			case 'external':
-				return new ExternalCircuitBreaker( $logSampler );
-			default:
-				return new NoopCircuitBreaker();
-		}
-	}
 }
